BỎ hết các thư viện, endponit tinh gọn, database, migrate, seeder, model, ngoài api chưa ok, login, và UI admin đồng bộ

+++ database/migrations/2025_10_27_194423_create_baiviet_table.php
@@ -19,7 +19,7 @@ class CreateBaivietTable extends Migration
             // DB::statement('ALTER TABLE baiviet MODIFY id INT(11) UNSIGNED AUTO_INCREMENT');
 
             // Cột 2: id_nguoidung (FK) - int(11) - Không NULL
-            $table->unsignedBigInteger('id_nguoidung');
+            $table->integer('id_nguoidung');
 
             // Cột 3: tieude - text - Không NULL
             $table->text('tieude');
